+++ RELEASENOTES.md
@@ -15,10 +15,14 @@
 ### [Datepickers](https://www.lightningdesignsystem.com/components/datepickers)
 #### Added
 - Added examples for Required, Required with Error states
+- Added `aria-label` to provide the full `DD MM YYYY` text of the date
+
 #### Fixed
 - Fixed missing attribute `aria-current="date"` when today's date falls inside of a selected range
 
 ### [Datetime Picker](https://www.lightningdesignsystem.com/components/datetime-picker)
+#### Added
+- Added `aria-label` to provide the full `DD MM YYYY` text of the date
 #### Changed
 - Removed the usage of the deprecated combobox. The Datetime Picker now uses the current combobox.